indexador/cortador/requirements.txt
pypdf>=3.0.0  
pillow>=9.0.0  
reportlab>=3.6.0  
pdf2image>=1.16.0